Reorganized buttons + misc visuals
