+++ src/components/Calendar/CalendarPaneToolbar.js
@@ -3,8 +3,10 @@ import { withStyles } from '@material-ui/core/styles';
 import { IconButton, Tooltip, Typography, InputBase } from '@material-ui/core';
 import { ChevronLeft, ChevronRight, Undo } from '@material-ui/icons';
 import PropTypes from 'prop-types';
-import Submenu from './Submenu';
+import CustomEventsDialog from '../CustomEvents/Popup';
 import DownloadMenu from './DownloadMenu';
+import FinalSwitch from './FinalSwitch';
+import ClearSchedButton from './ClearSchedButton';
 
 const styles = {
   toolbar: {
@@ -60,6 +62,61 @@ class CalendarPaneToolbar extends Component {
       });
     }
 
+    const events = this.props.eventsInCalendar;
+
+    let result = [];
+    let finalSchedule = [];
+    for (let item of events)
+      if (
+        !item.isCustomEvent &&
+        result.find(function(element) {
+          return element.courseCode === item.courseCode;
+        }) === undefined
+      )
+        result.push(item);
+
+    for (let course of result) {
+      if (course.section !== undefined) {
+        let final = course.section.finalExam;
+
+        if (final.length > 5) {
+          let [, , , date, start, startMin, end, endMin, ampm] = final.match(
+            /([A-za-z]+) *(\d{1,2}) *([A-za-z]+) *(\d{1,2}):(\d{2})-(\d{1,2}):(\d{2})(p?)/
+          );
+          start = parseInt(start, 10);
+          startMin = parseInt(startMin, 10);
+          end = parseInt(end, 10);
+          endMin = parseInt(endMin, 10);
+          date = [
+            date.includes('M'),
+            date.includes('Tu'),
+            date.includes('W'),
+            date.includes('Th'),
+            date.includes('F'),
+          ];
+          if (ampm === 'p' && end !== 12) {
+            start += 12;
+            end += 12;
+            if (start > end) start -= 12;
+          }
+
+          date.forEach((shouldBeInCal, index) => {
+            if (shouldBeInCal)
+              finalSchedule.push({
+                title: course.title,
+                courseType: 'Fin',
+                courseCode: course.courseCode,
+                location: course.location,
+                color: course.color,
+                isCustomEvent: false,
+                start: new Date(2018, 0, index + 1, start, startMin),
+                end: new Date(2018, 0, index + 1, end, endMin),
+              });
+          });
+        }
+      }
+    }
+
     return (
       <div className={classes.toolbar}>
         <IconButton onClick={() => this.props.onScheduleChange(0)}>
@@ -91,8 +148,16 @@ class CalendarPaneToolbar extends Component {
 
         <div className={classes.spacer} />
 
+        <FinalSwitch
+          displayFinal={this.props.displayFinal}
+          schedule={finalSchedule}
+          showFinalSchedule={this.props.showFinalSchedule}
+        />
+
+        <div className={classes.spacer} />
+
         <Tooltip title="Undo Last Delete">
-          <IconButton onClick={() => this.props.onUndo(null)}>
+          <IconButton onClick={() => this.props.onUndo(null)} fontSize="small">
             <Undo fontSize="small" />
           </IconButton>
         </Tooltip>
@@ -105,17 +170,16 @@ class CalendarPaneToolbar extends Component {
           />
         </Tooltip>
 
-        <Tooltip title="More">
-          <Submenu
-            onAddCustomEvent={this.props.onAddCustomEvent}
-            handleClearSchedule={this.props.handleClearSchedule}
-            onTakeScreenshot={this.props.onTakeScreenshot}
-            eventsInCalendar={this.props.eventsInCalendar}
-            showFinalSchedule={this.props.showFinalSchedule}
-            displayFinal={this.props.displayFinal}
-            currentScheduleIndex={this.props.currentScheduleIndex}
-          />
-        </Tooltip>
+        <CustomEventsDialog
+          onAddCustomEvent={this.props.onAddCustomEvent}
+          handleSubmenuClose={this.handleClose}
+        />
+
+        <ClearSchedButton
+          handleSubmenuClose={this.handleClose}
+          handleClearSchedule={this.props.handleClearSchedule}
+          currentScheduleIndex={this.props.currentScheduleIndex}
+        />
       </div>
     );
   }
