+++ src/components/Calendar/FinalSwitch.js
@@ -1,5 +1,5 @@
-import React from 'react';
-import { Switch, Typography } from '@material-ui/core';
+import React, { Fragment } from 'react';
+import { Switch } from '@material-ui/core';
 import FormControlLabel from '@material-ui/core/FormControlLabel';
 
 class FinalSwitch extends React.Component {
@@ -17,7 +17,7 @@ class FinalSwitch extends React.Component {
 
   render() {
     return (
-      <Typography>
+      <Fragment>
         <FormControlLabel
           control={
             <Switch
@@ -25,12 +25,11 @@ class FinalSwitch extends React.Component {
               onChange={this.handleChange('showFinal')}
               value="showFinal"
               color="primary"
-              style={{ margin: 0 }}
             />
           }
           label="FINALS"
         />
-      </Typography>
+      </Fragment>
     );
   }
 }
