+++ src/components/Calendar/calendar.css
@@ -44,3 +44,8 @@
 .menu-button:focus {
   background-color: unset !important;
 }
+
+.menu-button {
+  line-height: normal;
+  font-weight: 400;
+}
