+++ src/components/Map/Locator.js
@@ -3,6 +3,7 @@ import { withLeaflet } from 'react-leaflet';
 import Locate from 'leaflet.locatecontrol';
 
 class LocateControl extends Component {
+  // For geo-locating
   componentDidMount() {
     const { options } = this.props;
     const { map } = this.props.leaflet;
